move methods into rocblas.hpp
